refactor: 결과 그려주는 책임을 Output 으로 옮김
