+++ src/main/kotlin/racing/Main.kt
@@ -1,7 +1,6 @@
 package racing
 
 import racing.domain.RacingGame
-import racing.domain.ResultDto
 import racing.io.Input
 import racing.io.Output
 
@@ -12,5 +11,5 @@ fun main() {
     val racingGame = RacingGame(carCount, tryCount)
     val result = racingGame.play()
 
-    Output.resultMessage(ResultDto.from(result))
+    Output.resultMessage(result)
 }
